+++ numista_mobile/lib/services/wishlist_service.dart
@@ -1,5 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'auth_service.dart';
+import 'beta_checklist_service.dart';
 import '../models/coin_model.dart';
 
 class WishlistItem {
@@ -74,6 +75,7 @@ class WishlistService {
       'coin': coin.toFirestore(),
       'timestamp': FieldValue.serverTimestamp(),
     });
+    BetaChecklistService.autoCompleteTask('task_14_wishlist');
   }
 
   static Future<void> addProgramToWishlist(String programId, {
@@ -90,6 +92,7 @@ class WishlistService {
         'foundVarieties': initialVarieties,
         'timestamp': FieldValue.serverTimestamp(),
       });
+      BetaChecklistService.autoCompleteTask('task_14_wishlist');
     }
   }